refactor: 어색한 클래스명 정정
